+++ NExtra/Extensions/Object_PropertyExtensions.cs
@@ -1,0 +1,70 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq.Expressions;
+namespace NExtra.Extensions
+{
+    /// <summary>
+    /// Property-related extension methods for System.Object
+    /// </summary>
+    /// <remarks>
+    /// Author:     Daniel Saidi [daniel.saidi@gmail.com]
+    /// Link:       http://danielsaidi.github.com/nextra
+    /// 
+    /// The original, static implementation can be found at:
+    /// http://stackoverflow.com/questions/2789504/get-the-property-as-a-string-from-an-expressionfunctmodel-tproperty
+    /// </remarks>
+    public static class Object_PropertyExtensions
+    {
+        public static string GetFullPropertyName<T, TProperty>(this T obj, Expression<Func<T, TProperty>> expression)
+        {
+            return GetFullPropertyName(expression);
+        }
+
+
+        private static string GetFullPropertyName<T, TProperty>(Expression<Func<T, TProperty>> exp)
+        {
+            MemberExpression memberExp;
+            if (!TryFindMemberExpression(exp.Body, out memberExp))
+                return string.Empty;
+
+            var memberNames = new Stack<string>();
+            do { memberNames.Push(memberExp.Member.Name); }
+            while (TryFindMemberExpression(memberExp.Expression, out memberExp));
+
+            return string.Join(".", memberNames.ToArray());
+        }
+
+        private static bool TryFindMemberExpression(Expression exp, out MemberExpression memberExp)
+        {
+            memberExp = exp as MemberExpression;
+            if (memberExp != null)
+            {
+                // heyo! that was easy enough
+                return true;
+            }
+
+            // if the compiler created an automatic conversion, it'll look something like...
+            // obj => Convert(obj.Property) [e.g., int -> object]  OR:
+            // obj => ConvertChecked(obj.Property) [e.g., int -> long]
+            // ...which are the cases checked in IsConversion
+            if (IsConversion(exp) && exp is UnaryExpression)
+            {
+                memberExp = ((UnaryExpression)exp).Operand as MemberExpression;
+                if (memberExp != null)
+                {
+                    return true;
+                }
+            }
+
+            return false;
+        }
+
+        private static bool IsConversion(Expression exp)
+        {
+            return (
+                exp.NodeType == ExpressionType.Convert ||
+                exp.NodeType == ExpressionType.ConvertChecked
+            );
+        }
+    }
+}